requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
jmespath~=1.0.1
logging~=0.4.9.6
typing_extensions~=4.14.0
pillow~=11.3.0
h11~=0.16.0
python-dotenv~=1.1.1
mcp~=1.10.1
anyio~=4.9.0
pydantic~=2.11.7
httpx~=0.28.1
jsonschema~=4.23.0
starlette~=0.47.1
pydantic_core~=2.33.2
dill~=0.4.0
six~=1.17.0
colorama~=0.4.6
rich~=14.0.0
attrs~=24.2.0
Pygments~=2.19.2
zipp~=3.23.0
exceptiongroup~=1.3.0
sniffio~=1.3.1
botocore~=1.39.2
boto3~=1.39.2
s3transfer~=0.13.0
click~=8.1.7
httpcore~=1.0.9
idna~=3.10
mdurl~=0.1.2
slack_sdk~=3.35.0
sympy~=1.14.0
mpmath~=1.3.0
watchdog~=6.0.0
docstring_parser~=0.16
uvicorn~=0.35.0
wcwidth~=0.2.13
python-dateutil~=2.9.0.post0
urllib3~=2.5.0
certifi~=2025.6.15
tenacity~=9.1.2
itsdangerous~=2.2.0
Jinja2~=3.1.4
referencing~=0.35.1
slack_bolt~=1.23.0
Flask~=3.0.3
importlib_metadata~=8.7.0
prompt_toolkit~=3.0.51
requests~=2.32.4
python-multipart~=0.0.20
pip~=24.2
fastapi~=0.115.14